First4Aid_v3/requirements.txt
flask==2.1.2
pytest==7.1.2
flask-sqlalchemy==2.5.1
flask_bcrypt==1.0.1
requests==2.25.0
flask_wtf==1.0.1
wtforms==3.0.1
flask_login==0.6.2
wtforms_sqlalchemy==0.3
pyjwt==2.4.0
email-validator==1.2.1
Pillow==9.2.0
Flask-Mail==0.9.1